Added welcome, login, and signup screen
Using the React router, the welcome page links to the login page,
signup page, and the welcome page by clicking on links in the header.
The login page currently checks the input username and password in a
temporary database while the signup page checks that the input username
and email are not in the database.