Added Group and Transaction categories to database
Created new collections for MongoDB database and updated existing User
Collection. Commands still need to be tested and updated.

A Group document looks as such:

    Group {
        id,
	name,
	members (a list of User IDs),
	transactions (a list of Transaction IDs)
    }

A short description for the Group collection routes:

    List all groups: /group
    Retrieve group: /group/:group_id
    Create new group: /group/create
    Insert new user: /group/insert/user/:group_id
    Insert new transaction: /group/insert/transaction/:group_id
    Pop user: /group/pop/user/:group_id
    Pop transaction: /group/pop/transaction/:group_id
    Delete group: /group/delete/:group_id

A Transaction document looks as such:

    Transaction {
        id,
	loaner (a User ID),
	borrower (a User ID),
	amount
    }

A short description for the Transaction collection routes:

    List all transactions: /transaction
    Retrieve transaction: /transaction/:transaction_id
    Create new transaction: /transaction/create
    Update transaction amount: /transaction/update/:transaction_id
    Delete transaction: /transaction/delete/:transaction_id

These are not the final updates, User documents must be updated to allow
for storage of Groups and Friends, and extra update commands must be
added for all fields.