Added User Groups and Friends insertion/deletion
Added new routes to insert or remove Group and Friend IDs. Also fixed
bugs that wouldn't allow proper removal of existing entries in the Group
objects, as a new ID would be generated and used when searching for existing
IDs in the removal search query. Still need to add extra status messages
to be sent when a route function is called.