Add Transactions and Friends view to the Home page
Updated the homepage to list all transactions a user was a part of,
sorting them into transactions to pay off, and transactions that
need to be paid to the user.

Fixed minor errors in the creation of transactions, made sure to add
the transaction and group IDs to all relevant users.

Fixed issues with infinite rendering when trying to update database,
also made sure useEffect would be called at least twice, once to fetch
data and twice to render that data to the screen.