Add guards against accessing restricted pages
Added checks to make sure user was signed in if they tried to access
certain pages, otherwise alerts them that they must sign in and
redirects them to the Welcome page. Also switched out localStorage in
favor of sessionStorage so user login wouldn't persist if the tab was
closed or the app was reloaded.